requirements.txt
Flask==0.9
Flask-WTF==0.8
Jinja2==2.6
WTForms==1.0.2
Werkzeug==0.8.3
argparse==1.2.1
facebook-sdk==0.4.0
flask-mongoengine==0.6
gevent==0.13.8
greenlet==0.4.0
gunicorn==0.16.1
httplib2==0.7.7
mongoengine==0.7.6
newrelic==1.8.0.13
pymongo==2.3
wsgiref==0.1.2
boto==2.5.2